Extract out jwt auth code
